Dashboard section tab added to manage the admin controls
