+++ package.json
@@ -16,11 +16,13 @@
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/react-toast": "^1.2.6",
     "@types/node": "^22.13.5",
+    "bootstrap": "^5.3.7",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "lucide-react": "^0.475.0",
     "react": "18.3.1",
+    "react-bootstrap": "^2.10.10",
     "react-day-picker": "^8.10.1",
     "react-dom": "18.3.1",
     "react-router": "^7.2.0",
